refactor(lint): consolidate + relocate fixtures, document known aliasing gap
Proactive /simplify pass on PR #2129 (4 parallel review agents: reuse,
simplification, efficiency, altitude). Reuse and efficiency came back clean.
Two real, actionable findings applied:

- Simplification: 6 single-purpose fixture files had ~3 lines of duplicated
  import/header boilerplate each and some redundant in-file comments.
  Consolidated into 2 files grouped by scenario (db-handle-findmany-fixture.ts:
  unbounded/limited/findFirst on a plain `db` handle; alt-handle-and-nested-
  limit-findmany-fixture.ts: the `tx` handle case and the nested-limit case),
  each still independently exercising every case as its own exported function.

- Altitude: `src/app/api/__fixtures__/` sat inside the actual Next.js route
  tree. It happened to be silently skipped by two coverage-audit tests
  (drive-member-gate-coverage.test.ts, security-audit-coverage.test.ts) that
  filter by literal `route.ts` filename — but that's incidental, not a
  structural exclusion, and a future coverage walker that scans by extension
  instead would trip over it. Moved to src/__fixtures__/eslint-unbounded-
  findmany/ — still under `src/` (Next's default `next lint` scan dirs cover
  it, verified by temporarily adding an unsuppressed violation there and
  confirming `bun run lint --filter=web` failed on it, then reverting), but
  fully outside `app/api/`.

Also documented a known limitation directly in the rule's comment: it's a
syntactic AST pattern, not type-aware, so aliasing the query object first
(`const q = db.query.taskItems; q.findMany(...)`) evades it — the same class
of gap the tx/database fix (39532daa4) just closed for receiver identifiers.
Noted the durable fix would be a typed wrapper in packages/db, not another
AST special case, if this pattern shows up in practice.

Verified: bun run lint --filter=web and bun run typecheck --filter=web both
green after the move.
